pyproject.toml
[tool.poetry]
name = "template"
version = "0.1.2"
description = ""
authors = ["Uttam Dhabas <uttam@ariksa.com>"]

[tool.poetry.dependencies]
PyYAML = "5.3"
python = "^3.11"
sqlmodel = "^0.0.8"
asyncpg = "^0.28.0"
tenacity = "^8.0.1"
aioboto3 = "^10.0.0"
asyncio = "^3.4.3"
PyPDF2 = "^2.9.0"
pandas = "^1.4.3"
openpyxl = "^3.0.10"
cryptography = "^39.0.1"
pyOpenSSL = "^23.0.0"
snowflake-connector-python = "3.0.0"
asyncache = "^0.3.1"
apscheduler = "^3.9.1"
types-requests = "^2.28.11.13"
pymysql = "^1.0.2"
psutil = "^5.9.4"
psycopg2-binary = "^2.9.2"
xlrd = "^2.0.1"
loguru = "^0.7.0"
cassandra-driver = "^3.28.0"
python-dotenv = "0.21.1"
setuptools = "^67.8.0"
motor = "^3.2.0"
pillow = "9.5.0"
python-gitlab = "^3.15.0"
python-docx = "^1.1.0"
mitie = "^0.7.36"
google-re2 = "^1.1"
darker = "^1.7.1"
pympler = "^1.0.1"
memory-profiler = "^0.61.0"
hyperscan = "^0.7.7"

[tool.poetry.dev-dependencies]
httpx = "^0.23.0"
pip-audit = "^2.4.1"
pytest-asyncio = "^0.16.0"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-docker = "^0.10.3"
docker-compose = "^1.29.2"
pydantic-factories = "^1.0.0"

[tool.poetry.group.dev.dependencies]
darker = {extras = ["isort"], version = "^1.5.1"}

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
